refactor(marketplace): more typescript
